GEP-114 chore: Move restore logic to MinimizeHelper
